@@ -2,11 +2,8 @@
 /**
  * ITEA Office all rights reserved
  *
- * PHP Version 7
- *
- *
  * @author      Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright   Copyright (c) 2004-2018 ITEA Office (https://itea3.org)
+ * @copyright   Copyright (c) 2019 ITEA Office (https://itea3.org)
  * @license     https://itea3.org/license.txt proprietary
  *
  */
@@ -34,10 +31,7 @@ use ZfcTwig\View\TwigRenderer;
  */
 abstract class AbstractHandler extends AbstractHelper
 {
-    /**
-     * @var HelperPluginManager
-     */
-    protected $helperPluginManager;
+    protected HelperPluginManager $helperPluginManager;
     /**
      * @var RouteMatch
      */
@@ -45,7 +39,7 @@ abstract class AbstractHandler extends AbstractHelper
     /**
      * @var TwigRenderer
      */
-    protected $renderer;
+    protected TwigRenderer $renderer;
     /**
      * @var Response
      */
@@ -57,7 +51,7 @@ abstract class AbstractHandler extends AbstractHelper
     /**
      * @var TranslatorInterface
      */
-    protected $translator;
+    protected TranslatorInterface $translator;
 
     public function __construct(
         Application $application,
